refactor: move note extraction from advice provider into OutputNoteBuilder
